fix(build): correct bundle path for common and es modules
I didn't realize that the entire behavior of the typescript plugin for rollup had changed so
drastically in version 3. It used to produce a complete bundle, but now exports from other compiled
files, which means the path will be incorrect if the bundle isn't in the same directory. An
alternative to this approach would be to move the es and common module bundles into the lib
directory, but this would probably break typing, so it seems like the best approach is to keep the
compiled source at the root level. It seems simpler this way.