chore(package): add tslib peer dependency for bundlers
tslib is usually included with other packages, but in some uncommon cases, it may not be installed.
This will indicate to developers using this project that they may need to manually install tslib.
I've only encountered this when using rollup to bundle this project as an external dependency while
using the typescript plugin.